Moved Preference class into BraveShared module.
